+++ src/org/labkey/test/components/ui/grids/FieldSelectionDialog.java
@@ -7,7 +7,7 @@ import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.components.UpdatingComponent;
 import org.labkey.test.components.bootstrap.ModalDialog;
 import org.labkey.test.components.html.Checkbox;
-import org.labkey.test.util.selenium.WebDriverUtils;
+import org.labkey.test.util.selenium.WebElementUtils;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -21,7 +21,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static org.labkey.test.util.selenium.WebDriverUtils.getTextContent;
+import static org.labkey.test.util.selenium.WebElementUtils.getTextContent;
 
 /**
  * Wraps ColumnSelectionModal.tsx in UI components.
@@ -91,7 +91,7 @@ public class FieldSelectionDialog extends ModalDialog
     public List<String> getAvailableFieldLabels()
     {
         List<WebElement> listItemElements = elementCache().getListItemNameElements(elementCache().availableFieldsPanel);
-        return listItemElements.stream().map(WebDriverUtils::getTextContent).collect(Collectors.toList());
+        return listItemElements.stream().map(WebElementUtils::getTextContent).collect(Collectors.toList());
     }
 
     /**
@@ -243,7 +243,7 @@ public class FieldSelectionDialog extends ModalDialog
     public List<String> getSelectedFieldLabels()
     {
         List<WebElement> listItemElements = elementCache().getListItemNameElements(elementCache().selectedFieldsPanel);
-        return listItemElements.stream().map(WebDriverUtils::getTextContent).collect(Collectors.toList());
+        return listItemElements.stream().map(WebElementUtils::getTextContent).collect(Collectors.toList());
     }
 
     /**
